refactor(drinks): split DrinksScreen into focused widget files
drinks_screen.dart was the largest file in the app (1272 lines), mixing
the screen scaffold, festival header/banner, the bottom control buttons,
and four modal-sheet classes with some business logic inline. Split along
those seams, following the existing festival_menu_sheets.dart pattern, so
the screen drops to ~370 lines and each piece is individually testable.

- Extract the four filter sheets + show* helpers into
  widgets/drink_filter_sheets.dart; dedupe their drag handles.
- Extract the bottom-row buttons into widgets/drinks_filter_controls.dart.
- Extract the festival header into widgets/festival_header.dart and
  collapse the four parallel status-colour branches into a data-driven
  FestivalStatusBadge.
- Extract the festival banner into widgets/festival_banner.dart.
- Replace the private _sortLabel function with a DrinkSortLabel extension
  on DrinkSort.
- Move locale-aware style sorting out of the sheet build() into
  DrinkFilterController.availableStyles.
- Add tests for the new widgets, the sort labels, and the locale-aware
  available styles.

No behaviour change intended.

Closes #351
