+++ lib/screens/drinks_screen.dart
@@ -1,30 +1,11 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:go_router/go_router.dart';
 import 'package:provider/provider.dart';
 import '../domain/models/models.dart';
-import '../models/models.dart';
 import '../providers/providers.dart';
 import '../utils/utils.dart';
 import '../widgets/widgets.dart';
-
-String _sortLabel(DrinkSort sort) {
-  switch (sort) {
-    case DrinkSort.nameAsc:
-      return 'Name (A-Z)';
-    case DrinkSort.nameDesc:
-      return 'Name (Z-A)';
-    case DrinkSort.abvHigh:
-      return 'ABV (High to Low)';
-    case DrinkSort.abvLow:
-      return 'ABV (Low to High)';
-    case DrinkSort.brewery:
-      return 'Brewery (A-Z)';
-    case DrinkSort.style:
-      return 'Style (A-Z)';
-  }
-}
 
 /// Main screen showing the list of drinks
 class DrinksScreen extends StatefulWidget {
@@ -71,11 +52,14 @@ class _DrinksScreenState extends State<DrinksScreen> {
                   SliverAppBar(
                     floating: true,
                     snap: true,
-                    title: _buildFestivalHeader(context, provider),
+                    title: FestivalHeader(provider: provider),
                     actions: [buildOverflowMenu(context)],
                   ),
                   SliverToBoxAdapter(
-                    child: _buildFestivalBanner(context, provider),
+                    child: FestivalBanner(
+                      provider: provider,
+                      festivalId: widget.festivalId,
+                    ),
                   ),
                   SliverToBoxAdapter(
                     child: _buildRefreshStatus(context, provider),
@@ -152,50 +136,49 @@ class _DrinksScreenState extends State<DrinksScreen> {
       child: Row(
         children: [
           Expanded(
-            child: _FilterButton(
+            child: FilterButton(
               label: provider.selectedCategory ?? 'Category',
               semanticLabel: provider.selectedCategory != null
                   ? 'Filter by category: ${provider.selectedCategory}'
                   : 'Filter by category',
               icon: Icons.filter_list,
-              onPressed: () => _showCategoryFilter(context, provider),
+              onPressed: () => showCategoryFilter(context),
               isActive: provider.selectedCategory != null,
             ),
           ),
           if (hasStyleFilter) ...[
             const SizedBox(width: 6),
             Expanded(
-              child: _FilterButton(
+              child: FilterButton(
                 label: styleLabel,
                 semanticLabel: provider.selectedStyles.isEmpty
                     ? 'Filter by style'
                     : 'Filter by style: ${provider.selectedStyles.join(", ")}',
                 icon: Icons.style,
-                onPressed: () => _showStyleFilter(context, provider),
+                onPressed: () => showStyleFilter(context),
                 isActive: provider.selectedStyles.isNotEmpty,
               ),
             ),
           ],
           const SizedBox(width: 6),
           Expanded(
-            child: _FilterButton(
-              label: _sortLabel(provider.currentSort),
-              semanticLabel:
-                  'Sort drinks by ${_sortLabel(provider.currentSort)}',
+            child: FilterButton(
+              label: provider.currentSort.label,
+              semanticLabel: 'Sort drinks by ${provider.currentSort.label}',
               icon: Icons.sort,
-              onPressed: () => _showSortOptions(context, provider),
+              onPressed: () => showSortOptions(context),
               isActive: false,
             ),
           ),
           const SizedBox(width: 6),
-          _VisibilityFilterButton(
+          VisibilityFilterButton(
             activeCount:
                 provider.visibilityFilters.length +
                 provider.excludedAllergens.length,
-            onPressed: () => _showVisibilityFilter(context, provider),
+            onPressed: () => showVisibilityFilter(context),
           ),
           const SizedBox(width: 6),
-          _SearchButton(
+          SearchButton(
             isActive: _showSearch,
             hasQuery: provider.searchQuery.isNotEmpty,
             onPressed: () {
@@ -210,224 +193,6 @@ class _DrinksScreenState extends State<DrinksScreen> {
             },
           ),
         ],
-      ),
-    );
-  }
-
-  Widget _buildFestivalHeader(BuildContext context, BeerProvider provider) {
-    final theme = Theme.of(context);
-    final isDark = theme.brightness == Brightness.dark;
-    final status = Festival.getStatusInContext(
-      provider.currentFestival,
-      provider.sortedFestivals,
-    );
-
-    return Semantics(
-      label:
-          'Current festival: ${provider.currentFestival.name}, ${provider.drinks.length} drinks',
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Image.asset('assets/app_icon.png', width: 32, height: 32),
-          const SizedBox(width: 12),
-          Flexible(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                Text(
-                  provider.currentFestival.name,
-                  style: theme.textTheme.titleMedium?.copyWith(
-                    fontWeight: FontWeight.bold,
-                  ),
-                  overflow: TextOverflow.ellipsis,
-                  maxLines: 1,
-                ),
-                Row(
-                  children: [
-                    Flexible(
-                      child: Text(
-                        '${provider.drinks.length} drinks',
-                        style: theme.textTheme.bodySmall?.copyWith(
-                          color: theme.colorScheme.onSurfaceVariant,
-                        ),
-                        overflow: TextOverflow.ellipsis,
-                      ),
-                    ),
-                    if (status == FestivalStatus.live) ...[
-                      const SizedBox(width: 8),
-                      Container(
-                        padding: const EdgeInsets.symmetric(
-                          horizontal: 6,
-                          vertical: 1,
-                        ),
-                        decoration: BoxDecoration(
-                          color: isDark
-                              ? const Color(0xFF4CAF50)
-                              : const Color(0xFF2E7D32),
-                          borderRadius: BorderRadius.circular(8),
-                        ),
-                        child: const Text(
-                          'LIVE',
-                          style: TextStyle(
-                            color: Colors.white,
-                            fontSize: 9,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                      ),
-                    ] else if (status == FestivalStatus.upcoming) ...[
-                      const SizedBox(width: 8),
-                      Container(
-                        padding: const EdgeInsets.symmetric(
-                          horizontal: 6,
-                          vertical: 1,
-                        ),
-                        decoration: BoxDecoration(
-                          color: isDark
-                              ? const Color(0xFF42A5F5)
-                              : const Color(0xFF1976D2),
-                          borderRadius: BorderRadius.circular(8),
-                        ),
-                        child: const Text(
-                          'SOON',
-                          style: TextStyle(
-                            color: Colors.white,
-                            fontSize: 9,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                      ),
-                    ] else if (status == FestivalStatus.mostRecent) ...[
-                      const SizedBox(width: 8),
-                      Container(
-                        padding: const EdgeInsets.symmetric(
-                          horizontal: 6,
-                          vertical: 1,
-                        ),
-                        decoration: BoxDecoration(
-                          color: isDark
-                              ? const Color(0xFFFF9800)
-                              : const Color(0xFFEF6C00),
-                          borderRadius: BorderRadius.circular(8),
-                        ),
-                        child: const Text(
-                          'RECENT',
-                          style: TextStyle(
-                            color: Colors.white,
-                            fontSize: 9,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                      ),
-                    ] else if (status == FestivalStatus.past) ...[
-                      const SizedBox(width: 8),
-                      Container(
-                        padding: const EdgeInsets.symmetric(
-                          horizontal: 6,
-                          vertical: 1,
-                        ),
-                        decoration: BoxDecoration(
-                          color: isDark
-                              ? const Color(0xFF9E9E9E)
-                              : const Color(0xFF616161),
-                          borderRadius: BorderRadius.circular(8),
-                        ),
-                        child: const Text(
-                          'PAST',
-                          style: TextStyle(
-                            color: Colors.white,
-                            fontSize: 9,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                      ),
-                    ],
-                  ],
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildFestivalBanner(BuildContext context, BeerProvider provider) {
-    final theme = Theme.of(context);
-    final festival = provider.currentFestival;
-
-    // Only show banner if festival has dates or location
-    if (festival.formattedDates.isEmpty && festival.location == null) {
-      return const SizedBox.shrink();
-    }
-
-    final semanticLabel = [
-      if (festival.formattedDates.isNotEmpty) festival.formattedDates,
-      if (festival.location != null) festival.location,
-    ].join(', ');
-
-    return Material(
-      color: theme.colorScheme.primaryContainer.withValues(alpha: 0.3),
-      child: Semantics(
-        label: 'Festival information: $semanticLabel',
-        hint: 'Double tap for more details',
-        button: true,
-        child: InkWell(
-          onTap: () => context.push(buildFestivalInfoPath(widget.festivalId)),
-          child: Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-            child: Row(
-              children: [
-                Expanded(
-                  child: Wrap(
-                    spacing: 16,
-                    runSpacing: 4,
-                    children: [
-                      if (festival.formattedDates.isNotEmpty)
-                        Row(
-                          mainAxisSize: MainAxisSize.min,
-                          children: [
-                            Icon(
-                              Icons.calendar_today,
-                              size: 14,
-                              color: theme.colorScheme.onSurfaceVariant,
-                            ),
-                            const SizedBox(width: 4),
-                            Text(
-                              festival.formattedDates,
-                              style: theme.textTheme.bodySmall,
-                            ),
-                          ],
-                        ),
-                      if (festival.location != null)
-                        Row(
-                          mainAxisSize: MainAxisSize.min,
-                          children: [
-                            Icon(
-                              Icons.location_on,
-                              size: 14,
-                              color: theme.colorScheme.onSurfaceVariant,
-                            ),
-                            const SizedBox(width: 4),
-                            Text(
-                              festival.location!,
-                              style: theme.textTheme.bodySmall,
-                            ),
-                          ],
-                        ),
-                    ],
-                  ),
-                ),
-                Icon(
-                  Icons.chevron_right,
-                  size: 18,
-                  color: theme.colorScheme.onSurfaceVariant,
-                ),
-              ],
-            ),
-          ),
-        ),
       ),
     );
   }
@@ -602,670 +367,6 @@ class _DrinksScreenState extends State<DrinksScreen> {
     navigateToRoute(
       context,
       buildDrinkDetailPath(widget.festivalId, category, drinkId),
-    );
-  }
-
-  void _showCategoryFilter(BuildContext context, BeerProvider provider) {
-    showModalBottomSheet(
-      context: context,
-      isScrollControlled: true,
-      builder: (context) => _CategoryFilterSheet(provider: provider),
-    );
-  }
-
-  void _showStyleFilter(BuildContext context, BeerProvider provider) {
-    showModalBottomSheet(
-      context: context,
-      isScrollControlled: true,
-      builder: (context) => _StyleFilterSheet(provider: provider),
-    );
-  }
-
-  void _showSortOptions(BuildContext context, BeerProvider provider) {
-    showModalBottomSheet(
-      context: context,
-      isScrollControlled: true,
-      builder: (context) => _SortOptionsSheet(provider: provider),
-    );
-  }
-
-  void _showVisibilityFilter(BuildContext context, BeerProvider provider) {
-    showModalBottomSheet(
-      context: context,
-      isScrollControlled: true,
-      builder: (context) => const _VisibilityFilterSheet(),
-    );
-  }
-}
-
-class _FilterButton extends StatelessWidget {
-  final String label;
-  final IconData icon;
-  final VoidCallback onPressed;
-  final bool isActive;
-  final String? semanticLabel;
-
-  const _FilterButton({
-    required this.label,
-    required this.icon,
-    required this.onPressed,
-    required this.isActive,
-    this.semanticLabel,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final effectiveLabel = semanticLabel ?? label;
-    final semanticHint = isActive
-        ? 'Double tap to clear filter'
-        : 'Double tap to select filter';
-
-    return Semantics(
-      label: effectiveLabel,
-      hint: semanticHint,
-      button: true,
-      child: FilledButton.tonal(
-        onPressed: onPressed,
-        style: FilledButton.styleFrom(
-          padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
-        ),
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Icon(icon, size: 18),
-            const SizedBox(width: 8),
-            Flexible(
-              child: Text(
-                label,
-                overflow: TextOverflow.ellipsis,
-                style: theme.textTheme.labelMedium,
-              ),
-            ),
-            if (isActive) ...[
-              const SizedBox(width: 4),
-              const Icon(Icons.close, size: 16),
-            ],
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class _SearchButton extends StatelessWidget {
-  final bool isActive;
-  final bool hasQuery;
-  final VoidCallback onPressed;
-
-  const _SearchButton({
-    required this.isActive,
-    this.hasQuery = false,
-    required this.onPressed,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final label = isActive ? 'Close search' : 'Search drinks';
-    final hint = isActive
-        ? 'Double tap to close search bar'
-        : 'Double tap to open search bar';
-
-    return Semantics(
-      label: label,
-      hint: hint,
-      button: true,
-      child: FilledButton.tonal(
-        onPressed: onPressed,
-        style: FilledButton.styleFrom(
-          padding: const EdgeInsets.all(12),
-          minimumSize: const Size(48, 48),
-          backgroundColor: hasQuery && !isActive
-              ? theme.colorScheme.primaryContainer
-              : null,
-        ),
-        child: Icon(isActive ? Icons.search_off : Icons.search, size: 20),
-      ),
-    );
-  }
-}
-
-class _VisibilityFilterButton extends StatelessWidget {
-  final int activeCount;
-  final VoidCallback onPressed;
-
-  const _VisibilityFilterButton({
-    required this.activeCount,
-    required this.onPressed,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final isActive = activeCount > 0;
-    final label = isActive ? 'View filters ($activeCount)' : 'View filters';
-    const hint = 'Double tap to set availability and dietary filters';
-
-    return Semantics(
-      label: label,
-      hint: hint,
-      button: true,
-      child: FilledButton.tonal(
-        onPressed: onPressed,
-        style: FilledButton.styleFrom(
-          padding: const EdgeInsets.all(12),
-          minimumSize: const Size(48, 48),
-          backgroundColor: isActive ? theme.colorScheme.primaryContainer : null,
-        ),
-        child: Stack(
-          clipBehavior: Clip.none,
-          children: [
-            Icon(
-              isActive ? Icons.visibility : Icons.visibility_outlined,
-              size: 20,
-            ),
-            if (isActive)
-              Positioned(
-                right: -6,
-                top: -6,
-                child: Container(
-                  width: 14,
-                  height: 14,
-                  decoration: BoxDecoration(
-                    color: theme.colorScheme.primary,
-                    shape: BoxShape.circle,
-                  ),
-                  child: Center(
-                    child: Text(
-                      '$activeCount',
-                      style: TextStyle(
-                        color: theme.colorScheme.onPrimary,
-                        fontSize: 9,
-                        fontWeight: FontWeight.bold,
-                      ),
-                    ),
-                  ),
-                ),
-              ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class _CategoryFilterSheet extends StatelessWidget {
-  final BeerProvider provider;
-
-  const _CategoryFilterSheet({required this.provider});
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final categories = provider.availableCategories;
-    final counts = provider.categoryCountsMap;
-
-    return Container(
-      padding: const EdgeInsets.all(16),
-      constraints: BoxConstraints(
-        maxHeight: MediaQuery.of(context).size.height * 0.7,
-      ),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Center(
-            child: Container(
-              width: 32,
-              height: 4,
-              decoration: BoxDecoration(
-                color: theme.colorScheme.onSurfaceVariant,
-                borderRadius: BorderRadius.circular(2),
-              ),
-            ),
-          ),
-          const SizedBox(height: 16),
-          Text('Filter by Category', style: theme.textTheme.titleLarge),
-          const SizedBox(height: 16),
-          Flexible(
-            child: SingleChildScrollView(
-              child: RadioGroup<String?>(
-                groupValue: provider.selectedCategory,
-                onChanged: (value) {
-                  provider.setCategory(value);
-                  Navigator.pop(context);
-                },
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    Semantics(
-                      label:
-                          'Show all drinks, ${provider.allDrinks.length} total',
-                      selected: provider.selectedCategory == null,
-                      button: true,
-                      child: ListTile(
-                        leading: const Radio<String?>(value: null),
-                        title: Text('All (${provider.allDrinks.length})'),
-                        onTap: () {
-                          provider.setCategory(null);
-                          Navigator.pop(context);
-                        },
-                      ),
-                    ),
-                    ...categories.map((category) {
-                      final formattedCategory =
-                          BeverageTypeHelper.formatBeverageType(category);
-                      final count = counts[category] ?? 0;
-                      return Semantics(
-                        label: 'Filter by $formattedCategory, $count drinks',
-                        selected: provider.selectedCategory == category,
-                        button: true,
-                        child: ListTile(
-                          leading: Radio<String?>(value: category),
-                          title: Text('$formattedCategory ($count)'),
-                          onTap: () {
-                            provider.setCategory(category);
-                            Navigator.pop(context);
-                          },
-                        ),
-                      );
-                    }),
-                  ],
-                ),
-              ),
-            ),
-          ),
-          const SizedBox(height: 16),
-        ],
-      ),
-    );
-  }
-}
-
-class _SortOptionsSheet extends StatelessWidget {
-  final BeerProvider provider;
-
-  const _SortOptionsSheet({required this.provider});
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-
-    return Container(
-      padding: const EdgeInsets.all(16),
-      constraints: BoxConstraints(
-        maxHeight: MediaQuery.of(context).size.height * 0.7,
-      ),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Center(
-            child: Container(
-              width: 32,
-              height: 4,
-              decoration: BoxDecoration(
-                color: theme.colorScheme.onSurfaceVariant,
-                borderRadius: BorderRadius.circular(2),
-              ),
-            ),
-          ),
-          const SizedBox(height: 16),
-          Text('Sort By', style: theme.textTheme.titleLarge),
-          const SizedBox(height: 16),
-          Flexible(
-            child: SingleChildScrollView(
-              child: RadioGroup<DrinkSort>(
-                groupValue: provider.currentSort,
-                onChanged: (value) {
-                  if (value != null) {
-                    provider.setSort(value);
-                    Navigator.pop(context);
-                  }
-                },
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  children: DrinkSort.values.map((sort) {
-                    final sortLabel = _sortLabel(sort);
-                    return Semantics(
-                      label: 'Sort by $sortLabel',
-                      selected: provider.currentSort == sort,
-                      button: true,
-                      child: ListTile(
-                        leading: Radio<DrinkSort>(value: sort),
-                        title: Text(sortLabel),
-                        onTap: () {
-                          provider.setSort(sort);
-                          Navigator.pop(context);
-                        },
-                      ),
-                    );
-                  }).toList(),
-                ),
-              ),
-            ),
-          ),
-          const SizedBox(height: 16),
-        ],
-      ),
-    );
-  }
-}
-
-/// Style filter sheet with checkboxes for multi-select
-class _StyleFilterSheet extends StatelessWidget {
-  final BeerProvider provider;
-
-  const _StyleFilterSheet({required this.provider});
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-
-    return Consumer<BeerProvider>(
-      builder: (context, beerProvider, child) {
-        final styles = beerProvider.availableStyles;
-        final styleCounts = beerProvider.styleCountsMap;
-        final selectedStyles = beerProvider.selectedStyles;
-
-        // Sort styles alphabetically using locale-aware comparison
-        // This ensures non-ASCII characters (é, ñ, etc.) sort correctly
-        final sortedStyles = List<String>.from(styles);
-        sortedStyles.sort(StringComparisonHelper.compareLocaleAware);
-
-        return Container(
-          padding: const EdgeInsets.all(16),
-          constraints: BoxConstraints(
-            maxHeight: MediaQuery.of(context).size.height * 0.7,
-          ),
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Center(
-                child: Container(
-                  width: 32,
-                  height: 4,
-                  decoration: BoxDecoration(
-                    color: theme.colorScheme.onSurfaceVariant,
-                    borderRadius: BorderRadius.circular(2),
-                  ),
-                ),
-              ),
-              const SizedBox(height: 16),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  Text('Filter by Style', style: theme.textTheme.titleLarge),
-                  if (selectedStyles.isNotEmpty)
-                    Semantics(
-                      label: 'Clear all style filters',
-                      hint: 'Double tap to remove all style filters',
-                      button: true,
-                      child: TextButton.icon(
-                        icon: const Icon(Icons.clear, size: 18),
-                        label: const Text('Clear'),
-                        onPressed: () {
-                          beerProvider.clearStyles();
-                        },
-                        style: TextButton.styleFrom(
-                          visualDensity: VisualDensity.compact,
-                        ),
-                      ),
-                    ),
-                ],
-              ),
-              const SizedBox(height: 8),
-              // Always reserve space for selected styles to prevent list jumping
-              AnimatedSize(
-                duration: const Duration(milliseconds: 200),
-                curve: Curves.easeInOut,
-                child: selectedStyles.isNotEmpty
-                    ? Container(
-                        padding: const EdgeInsets.symmetric(
-                          horizontal: 12,
-                          vertical: 8,
-                        ),
-                        decoration: BoxDecoration(
-                          color: theme.colorScheme.primaryContainer,
-                          borderRadius: BorderRadius.circular(8),
-                        ),
-                        child: Row(
-                          children: [
-                            Icon(
-                              Icons.check_circle,
-                              size: 16,
-                              color: theme.colorScheme.onPrimaryContainer,
-                            ),
-                            const SizedBox(width: 8),
-                            Expanded(
-                              child: Text(
-                                selectedStyles.join(', '),
-                                style: theme.textTheme.bodyMedium?.copyWith(
-                                  color: theme.colorScheme.onPrimaryContainer,
-                                ),
-                                maxLines: 2,
-                                overflow: TextOverflow.ellipsis,
-                              ),
-                            ),
-                          ],
-                        ),
-                      )
-                    : const SizedBox.shrink(),
-              ),
-              const SizedBox(height: 16),
-              Flexible(
-                child: SingleChildScrollView(
-                  child: Column(
-                    mainAxisSize: MainAxisSize.min,
-                    children: sortedStyles.map((style) {
-                      final count = styleCounts[style] ?? 0;
-                      final isSelected = selectedStyles.contains(style);
-                      return Semantics(
-                        label: 'Filter by $style, $count drinks',
-                        value: isSelected ? 'Selected' : 'Not selected',
-                        selected: isSelected,
-                        button: true,
-                        child: CheckboxListTile(
-                          value: isSelected,
-                          onChanged: (_) => beerProvider.toggleStyle(style),
-                          title: Text('$style ($count)'),
-                          controlAffinity: ListTileControlAffinity.leading,
-                          dense: true,
-                        ),
-                      );
-                    }).toList(),
-                  ),
-                ),
-              ),
-              const SizedBox(height: 16),
-            ],
-          ),
-        );
-      },
-    );
-  }
-}
-
-/// Sheet for toggling visibility filters (availability, tasted, vegan, allergen-free)
-class _VisibilityFilterSheet extends StatelessWidget {
-  const _VisibilityFilterSheet();
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-
-    return Consumer<BeerProvider>(
-      builder: (context, beerProvider, child) {
-        final active = beerProvider.visibilityFilters;
-
-        return Container(
-          padding: const EdgeInsets.all(16),
-          constraints: BoxConstraints(
-            maxHeight: MediaQuery.of(context).size.height * 0.7,
-          ),
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Center(
-                child: Container(
-                  width: 32,
-                  height: 4,
-                  decoration: BoxDecoration(
-                    color: theme.colorScheme.onSurfaceVariant,
-                    borderRadius: BorderRadius.circular(2),
-                  ),
-                ),
-              ),
-              const SizedBox(height: 16),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  Text('View Filters', style: theme.textTheme.titleLarge),
-                  if (active.isNotEmpty ||
-                      beerProvider.excludedAllergens.isNotEmpty)
-                    Semantics(
-                      label: 'Clear all view filters',
-                      hint: 'Double tap to remove all view filters',
-                      button: true,
-                      child: TextButton.icon(
-                        icon: const Icon(Icons.clear, size: 18),
-                        label: const Text('Clear'),
-                        onPressed: () async {
-                          await beerProvider.clearVisibilityFilters();
-                          await beerProvider.clearAllergenFilters();
-                        },
-                        style: TextButton.styleFrom(
-                          visualDensity: VisualDensity.compact,
-                        ),
-                      ),
-                    ),
-                ],
-              ),
-              const SizedBox(height: 8),
-              Flexible(
-                child: SingleChildScrollView(
-                  child: Column(
-                    mainAxisSize: MainAxisSize.min,
-                    children: [
-                      _VisibilityFilterTile(
-                        label: 'Available only',
-                        subtitle: 'Hide sold out & not yet arrived drinks',
-                        icon: Icons.check_circle_outline,
-                        isChecked: active.contains(
-                          DrinkVisibilityFilter.availableOnly,
-                        ),
-                        onChanged: (value) => beerProvider.setVisibilityFilter(
-                          DrinkVisibilityFilter.availableOnly,
-                          value ?? false,
-                        ),
-                      ),
-                      _VisibilityFilterTile(
-                        label: 'Not tasted',
-                        subtitle: 'Hide drinks you\'ve already tasted',
-                        icon: Icons.remove_circle_outline,
-                        isChecked: active.contains(
-                          DrinkVisibilityFilter.notTasted,
-                        ),
-                        onChanged: (value) => beerProvider.setVisibilityFilter(
-                          DrinkVisibilityFilter.notTasted,
-                          value ?? false,
-                        ),
-                      ),
-                      _VisibilityFilterTile(
-                        label: 'Vegan only',
-                        subtitle: 'Show only drinks marked as vegan',
-                        icon: Icons.eco_outlined,
-                        isChecked: active.contains(
-                          DrinkVisibilityFilter.veganOnly,
-                        ),
-                        onChanged: (value) => beerProvider.setVisibilityFilter(
-                          DrinkVisibilityFilter.veganOnly,
-                          value ?? false,
-                        ),
-                      ),
-                      if (beerProvider.availableAllergens.isNotEmpty) ...[
-                        const Divider(),
-                        Padding(
-                          padding: const EdgeInsets.symmetric(
-                            horizontal: 16,
-                            vertical: 4,
-                          ),
-                          child: Text(
-                            'Allergen-free',
-                            style: Theme.of(context).textTheme.labelMedium
-                                ?.copyWith(
-                                  color: Theme.of(
-                                    context,
-                                  ).colorScheme.onSurfaceVariant,
-                                ),
-                          ),
-                        ),
-                        for (final allergen
-                            in (beerProvider.availableAllergens.toList()
-                              ..sort()))
-                          _VisibilityFilterTile(
-                            label: _formatAllergenName(allergen),
-                            subtitle: 'Hide drinks containing $allergen',
-                            icon: Icons.no_meals_outlined,
-                            isChecked: beerProvider.excludedAllergens.contains(
-                              allergen,
-                            ),
-                            onChanged: (value) => beerProvider
-                                .setAllergenFilter(allergen, value ?? false),
-                          ),
-                      ],
-                    ],
-                  ),
-                ),
-              ),
-              const SizedBox(height: 16),
-            ],
-          ),
-        );
-      },
-    );
-  }
-
-  static String _formatAllergenName(String allergen) {
-    if (allergen.isEmpty) return allergen;
-    return allergen[0].toUpperCase() + allergen.substring(1);
-  }
-}
-
-class _VisibilityFilterTile extends StatelessWidget {
-  final String label;
-  final String subtitle;
-  final IconData icon;
-  final bool isChecked;
-  final ValueChanged<bool?> onChanged;
-
-  const _VisibilityFilterTile({
-    required this.label,
-    required this.subtitle,
-    required this.icon,
-    required this.isChecked,
-    required this.onChanged,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Semantics(
-      label: '$label. $subtitle',
-      value: isChecked ? 'Active' : 'Inactive',
-      selected: isChecked,
-      button: true,
-      child: CheckboxListTile(
-        value: isChecked,
-        onChanged: onChanged,
-        secondary: Icon(icon),
-        title: Text(label),
-        subtitle: Text(subtitle),
-        controlAffinity: ListTileControlAffinity.leading,
-        dense: true,
-      ),
     );
   }
 }
