+++ lib/widgets/festival_header.dart
@@ -1,0 +1,107 @@
+import 'package:flutter/material.dart';
+import '../models/models.dart';
+import '../providers/providers.dart';
+
+/// App-bar title for the drinks screen: app icon, current festival name, the
+/// drink count, and a coloured status badge.
+class FestivalHeader extends StatelessWidget {
+  const FestivalHeader({required this.provider, super.key});
+
+  final BeerProvider provider;
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    final status = Festival.getStatusInContext(
+      provider.currentFestival,
+      provider.sortedFestivals,
+    );
+
+    return Semantics(
+      label:
+          'Current festival: ${provider.currentFestival.name}, ${provider.drinks.length} drinks',
+      child: Row(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          Image.asset('assets/app_icon.png', width: 32, height: 32),
+          const SizedBox(width: 12),
+          Flexible(
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                Text(
+                  provider.currentFestival.name,
+                  style: theme.textTheme.titleMedium?.copyWith(
+                    fontWeight: FontWeight.bold,
+                  ),
+                  overflow: TextOverflow.ellipsis,
+                  maxLines: 1,
+                ),
+                Row(
+                  children: [
+                    Flexible(
+                      child: Text(
+                        '${provider.drinks.length} drinks',
+                        style: theme.textTheme.bodySmall?.copyWith(
+                          color: theme.colorScheme.onSurfaceVariant,
+                        ),
+                        overflow: TextOverflow.ellipsis,
+                      ),
+                    ),
+                    const SizedBox(width: 8),
+                    FestivalStatusBadge(status: status),
+                  ],
+                ),
+              ],
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+/// Small coloured pill summarising a festival's [FestivalStatus]
+/// (LIVE / SOON / RECENT / PAST). Colours adapt to light and dark themes.
+class FestivalStatusBadge extends StatelessWidget {
+  const FestivalStatusBadge({required this.status, super.key});
+
+  final FestivalStatus status;
+
+  @override
+  Widget build(BuildContext context) {
+    final isDark = Theme.of(context).brightness == Brightness.dark;
+    final (label, lightColor, darkColor) = _styleFor(status);
+
+    return Container(
+      padding: const EdgeInsets.symmetric(horizontal: 6, vertical: 1),
+      decoration: BoxDecoration(
+        color: isDark ? darkColor : lightColor,
+        borderRadius: BorderRadius.circular(8),
+      ),
+      child: Text(
+        label,
+        style: const TextStyle(
+          color: Colors.white,
+          fontSize: 9,
+          fontWeight: FontWeight.bold,
+        ),
+      ),
+    );
+  }
+
+  /// Returns the badge label and its (light, dark) background colours.
+  static (String, Color, Color) _styleFor(FestivalStatus status) {
+    switch (status) {
+      case FestivalStatus.live:
+        return const ('LIVE', Color(0xFF2E7D32), Color(0xFF4CAF50));
+      case FestivalStatus.upcoming:
+        return const ('SOON', Color(0xFF1976D2), Color(0xFF42A5F5));
+      case FestivalStatus.mostRecent:
+        return const ('RECENT', Color(0xFFEF6C00), Color(0xFFFF9800));
+      case FestivalStatus.past:
+        return const ('PAST', Color(0xFF616161), Color(0xFF9E9E9E));
+    }
+  }
+}
