+++ lib/widgets/widgets.dart
@@ -3,8 +3,12 @@ export 'bottom_action_bar.dart';
 export 'breadcrumb_bar.dart';
 export 'detail_header.dart';
 export 'drink_card.dart';
+export 'drink_filter_sheets.dart';
 export 'drink_list_section.dart';
+export 'drinks_filter_controls.dart';
 export 'environment_badge.dart';
+export 'festival_banner.dart';
+export 'festival_header.dart';
 export 'festival_menu_sheets.dart';
 export 'hero_info_card.dart';
 export 'info_chip.dart';
